fixed error screen showing up bug
error screen showed up when it shouldn't because the game error property would be set to true if the game a user was listening to was deleted. game fetch would be nil and game error property would be set to true in the listener. fixed behavior by commenting out error var being set to true here since the user would repopulate this with a non nil value when joining a new game.